Feat: main viewmodel 공유 로직 추가
