+++ app/src/main/java/com/asap/aljyo/core/components/viewmodel/main/MainViewModel.kt
@@ -1,5 +1,6 @@
 package com.asap.aljyo.core.components.viewmodel.main
 
+import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -12,6 +13,10 @@ import javax.inject.Inject
 class MainViewModel @Inject constructor() : ViewModel() {
     private val _selectedIndex = MutableStateFlow(0)
     val selectedIndex get() = _selectedIndex.asStateFlow()
+
+    init {
+        Log.d("MainViewModel", "main view model init")
+    }
 
     fun select(index: Int) = viewModelScope.launch {
         _selectedIndex.emit(index)
